extend caching ttl considerations
Cache-Control is not the only header that can determine how long a
response is considered valid. Expires is old but still widely in use. If
cache-control is not present, Expires should be considered next (if
available) to determine a ttl.

I purposely did not use Utils::getMaxAge() since it does not take into
account max-stale and I was not sure if that function should be modified
or not (with concerns to backwards compatibility etc).